Add support for WordPress Multisite Installations
By default, Bugsnag requires a separate API key configuration for each
site in a multisite network. For most networks, this is not ideal. All
sites in the network should report to a single Bugsnag dashboard. If
Bugsnag is Network Activated in a Multisite installation, it will now
place the options page in the Network admin panel and only require one
API key (and other settings) for the whole network. If the plugin is
enabled on a per-site basis, it works the same as before, where each
site can have its own settings.